Created Rocket Movement Component
This custom component based on projectile movement component makes sure that our rocket continues to move even if the projectile movement component registers a hit. We only want to register hits with our collision box on the rocket class. Now that this is functioning, we will not blow ourselves up on occasion while running the same direction as the rocket is flying.