Scatter Finished
Shotgun uses a TMap to map blaster character pointers to ints that represent how many pellets hit the character to tally damage for one call to apply damage. This is nice because it also allows the tally of damage across multiple characters.

Known Bugs:
Server Health Max is changed when shot -- Ex 56/56 instead of 56/100

Restart Not working